+++ doc/changelog.md
@@ -149,8 +149,10 @@ v0.6.6
 + (done, tested) Widget rendering is now doesn't depend on session, ommiting possibility of clutter display
 + (done, tested) Per subsite module & theme configuration 
 + (done, tested) Subsite module & theme privilege mechanism
-+ (proposed) nordrassil priority index and table prefix
-+ (proposed) namespace
++ (done, tested) nordrassil priority index and table prefix
 + (proposed) IP Deny manager.
+
+v0.6.7
++ (proposed) namespace
 + (proposed) Navigation: move to anywhere.
 + (proposed) Wait for the new version of CI 3.0 or apply this fix to Codeigniter development team: select_max does not escape field_name, pgsql driver should convert `DATETIME` to `TIMESTAMP` since there is no `DATETIME` in postgre <-- The first one seems to be fixed